fix week day list issues raised by daylight saving
this is an edge issue when we just entered dst. take 2014-03-11 for example,
if TimeTrackingFirstDayOfWeek is set to "Wednesday" then the week will
be 2014-03-05 -- 2014-03-11. without this commit, the week days are actaully
2014-03-04 -- 2014-03-08, and then jumps to 2014-03-10, 2014-03-11.

this is because when we call ->AddDays, we simply add seconds(24*3600)
actually instead of days, which is not right when dst is involved.