use different UpdateTimeWorked fields for tickets in one form
in 9820915, the change of "Move per-ticket save button to a single day button"
results in multiple id/UpdateTimeWorked inputs in one form when there are
multiple tickets on the day, sadly the process code hasn't been updated.
(so there will be an error of "Could not load ticket ARRAY(0x8118bb030)")

let's use a unique input name for each ticket to make it more clearly.